refactor: reduce tool complexity
