+++ src/tools/project_index.rs
@@ -2,10 +2,15 @@ use std::{fs, path::Path, sync::Arc, thread};
 
 use serde::{Deserialize, Serialize};
 
+use self::project_index_scan::scan_text_file;
 use super::ScanRoot;
-use super::hoi4_keys::flag_entity_type;
-use super::paradox_lexer::{Token, TokenKind, tokenize};
 use super::project_files::{ProjectFile, collect_project_files};
+
+#[path = "project_index_scan.rs"]
+mod project_index_scan;
+#[cfg(test)]
+#[path = "project_index_tests.rs"]
+mod project_index_tests;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub struct ProjectIndexRequest {
@@ -44,20 +49,20 @@ pub struct ProjectIndexResult {
 }
 
 #[derive(Debug, Clone)]
-struct ScanFile {
-    root: String,
-    root_role: Option<String>,
-    absolute_path: std::path::PathBuf,
-    relative_path: String,
-    file_type: String,
-    bytes: u64,
+pub(super) struct ScanFile {
+    pub(super) root: String,
+    pub(super) root_role: Option<String>,
+    pub(super) absolute_path: std::path::PathBuf,
+    pub(super) relative_path: String,
+    pub(super) file_type: String,
+    pub(super) bytes: u64,
 }
 
 #[derive(Debug, Clone, Default)]
-struct WorkerOutput {
-    files: Vec<IndexedFile>,
-    definitions: Vec<ProjectIndexItem>,
-    references: Vec<ProjectIndexItem>,
+pub(super) struct WorkerOutput {
+    pub(super) files: Vec<IndexedFile>,
+    pub(super) definitions: Vec<ProjectIndexItem>,
+    pub(super) references: Vec<ProjectIndexItem>,
 }
 
 pub fn index_hoi4_project(request: ProjectIndexRequest) -> Result<ProjectIndexResult, String> {
@@ -171,335 +176,6 @@ fn scan_file_chunk(files: &[ScanFile]) -> WorkerOutput {
     output
 }
 
-fn scan_text_file(file: &ScanFile, content: &str, output: &mut WorkerOutput) {
-    scan_localisation(file, content, output);
-
-    let tokens = tokenize(content);
-    let mut stack = Vec::<String>::new();
-    let mut index = 0usize;
-
-    while index < tokens.len() {
-        let token = &tokens[index];
-
-        if token.kind == TokenKind::Close {
-            stack.pop();
-            index += 1;
-            continue;
-        }
-
-        if is_block_start(&tokens, index) {
-            let key = tokens[index].text.clone();
-            scan_block_definition(file, &key, token.line, &stack, output);
-            stack.push(key);
-            index += 3;
-            continue;
-        }
-
-        if is_assignment(&tokens, index) {
-            let key = &tokens[index].text;
-            let value = &tokens[index + 2].text;
-            scan_assignment(file, key, value, token.line, &stack, output);
-            index += 3;
-            continue;
-        }
-
-        index += 1;
-    }
-}
-
-fn scan_block_definition(
-    file: &ScanFile,
-    key: &str,
-    line: usize,
-    stack: &[String],
-    output: &mut WorkerOutput,
-) {
-    let path = file.relative_path.as_str();
-    let parent = stack.last().map(String::as_str);
-
-    if let Some((kind, context)) = top_level_scripted_kind(path, stack.is_empty()) {
-        push_definition(file, output, kind, key, line, context);
-    }
-
-    if idea_token_definition(path, key) {
-        push_definition(file, output, "idea_token", key, line, "idea token block");
-    }
-
-    if dynamic_modifier_definition(path, parent) {
-        push_definition(
-            file,
-            output,
-            "dynamic_modifier",
-            key,
-            line,
-            "dynamic modifier block",
-        );
-    }
-}
-
-fn top_level_scripted_kind(path: &str, is_top_level: bool) -> Option<(&'static str, &'static str)> {
-    if !is_top_level {
-        return None;
-    }
-    if path.starts_with("common/scripted_triggers/") {
-        Some(("scripted_trigger", "top-level scripted trigger"))
-    } else if path.starts_with("common/scripted_effects/") {
-        Some(("scripted_effect", "top-level scripted effect"))
-    } else {
-        None
-    }
-}
-
-fn idea_token_definition(path: &str, key: &str) -> bool {
-    path.starts_with("common/ideas/") && !is_ignored_idea_block(key)
-}
-
-fn dynamic_modifier_definition(path: &str, parent: Option<&str>) -> bool {
-    path.starts_with("common/dynamic_modifiers/") || parent == Some("dynamic_modifier")
-}
-
-fn scan_assignment(
-    file: &ScanFile,
-    key: &str,
-    value: &str,
-    line: usize,
-    stack: &[String],
-    output: &mut WorkerOutput,
-) {
-    let current_block = stack.last().map(String::as_str);
-
-    scan_asset_assignment(file, key, value, line, current_block, output);
-    scan_flag_assignment(file, key, value, line, current_block, output);
-    scan_variable_assignment(file, key, value, line, current_block, output);
-    scan_focus_event_assignment(file, key, value, line, current_block, output);
-    scan_country_tag_assignment(file, key, line, output);
-}
-
-fn scan_asset_assignment(
-    file: &ScanFile,
-    key: &str,
-    value: &str,
-    line: usize,
-    current_block: Option<&str>,
-    output: &mut WorkerOutput,
-) {
-    if let Some((kind, context)) = asset_definition_kind(file, key, current_block) {
-        push_definition(file, output, kind, value, line, context);
-    }
-
-    if let Some((kind, context)) = asset_reference_kind(key, current_block) {
-        push_reference(file, output, kind, value, line, context);
-    }
-}
-
-fn asset_definition_kind<'a>(
-    file: &ScanFile,
-    key: &str,
-    current_block: Option<&'a str>,
-) -> Option<(&'static str, &'a str)> {
-    if key != "name" {
-        return None;
-    }
-    match current_block {
-        Some("spriteType") => Some(("gfx_sprite", "spriteType name")),
-        Some(block)
-            if is_gui_element_block(block) && file.relative_path.starts_with("interface/") =>
-        {
-            Some(("gui_element", block))
-        }
-        _ => None,
-    }
-}
-
-fn asset_reference_kind<'a>(
-    key: &'a str,
-    current_block: Option<&str>,
-) -> Option<(&'static str, &'a str)> {
-    match key {
-        "texturefile" if current_block == Some("spriteType") => {
-            Some(("asset_texture", "sprite texturefile"))
-        }
-        "quadTextureSprite" | "spriteType" => Some(("gfx_sprite", key)),
-        _ => None,
-    }
-}
-
-fn is_gui_element_block(block: &str) -> bool {
-    matches!(
-        block,
-        "containerWindowType" | "buttonType" | "iconType" | "instantTextBoxType" | "listboxType"
-    )
-}
-
-fn scan_flag_assignment(
-    file: &ScanFile,
-    key: &str,
-    value: &str,
-    line: usize,
-    current_block: Option<&str>,
-    output: &mut WorkerOutput,
-) {
-    if let Some(flag_kind) = flag_entity_type(key) {
-        push_reference(file, output, flag_kind, value, line, key);
-    }
-
-    if key == "flag"
-        && let Some(block) = current_block.and_then(flag_entity_type)
-    {
-        push_reference(file, output, block, value, line, "flag field");
-    }
-}
-
-fn scan_variable_assignment(
-    file: &ScanFile,
-    key: &str,
-    value: &str,
-    line: usize,
-    current_block: Option<&str>,
-    output: &mut WorkerOutput,
-) {
-    if is_variable_key(key) {
-        push_reference(file, output, "variable", value, line, key);
-    }
-
-    if current_block.is_some_and(is_variable_key)
-        && let Some(variable_name) = variable_name_from_field(key, value)
-    {
-        push_reference(
-            file,
-            output,
-            "variable",
-            variable_name,
-            line,
-            "variable field",
-        );
-    }
-}
-
-fn scan_focus_event_assignment(
-    file: &ScanFile,
-    key: &str,
-    value: &str,
-    line: usize,
-    current_block: Option<&str>,
-    output: &mut WorkerOutput,
-) {
-    if let Some((kind, context)) = id_definition_kind(key, current_block) {
-        push_definition(file, output, kind, value, line, context);
-    }
-
-    if matches!(key, "shared_focus" | "joint_focus") {
-        push_reference(file, output, "focus_id", value, line, key);
-    }
-
-    if key == "namespace" {
-        push_definition(
-            file,
-            output,
-            "event_namespace",
-            value,
-            line,
-            "event namespace",
-        );
-    }
-}
-
-fn id_definition_kind(
-    key: &str,
-    current_block: Option<&str>,
-) -> Option<(&'static str, &'static str)> {
-    if key != "id" {
-        return None;
-    }
-    match current_block {
-        Some("focus" | "shared_focus" | "joint_focus") => Some(("focus_id", "focus id")),
-        Some("focus_tree") => Some(("focus_tree_id", "focus tree id")),
-        Some("country_event" | "news_event" | "state_event" | "unit_event") => {
-            Some(("event_id", "event id"))
-        }
-        _ => None,
-    }
-}
-
-fn scan_country_tag_assignment(file: &ScanFile, key: &str, line: usize, output: &mut WorkerOutput) {
-    let path = file.relative_path.as_str();
-    if path.starts_with("common/country_tags/") {
-        push_definition(file, output, "country_tag", key, line, "country tag");
-    }
-}
-
-fn scan_localisation(file: &ScanFile, content: &str, output: &mut WorkerOutput) {
-    if !file.relative_path.starts_with("localisation/") {
-        return;
-    }
-
-    for (line_index, line) in content.lines().enumerate() {
-        let trimmed = line.trim_start().trim_start_matches('\u{feff}');
-        let Some((key, rest)) = trimmed.split_once(':') else {
-            continue;
-        };
-
-        let key = key.trim();
-        let rest = rest.trim_start();
-        if key.is_empty() || is_localisation_language_header(key, rest) {
-            continue;
-        }
-
-        push_definition(
-            file,
-            output,
-            "localisation_key",
-            key,
-            line_index + 1,
-            "localisation key",
-        );
-    }
-}
-
-fn push_definition(
-    file: &ScanFile,
-    output: &mut WorkerOutput,
-    kind: &str,
-    name: &str,
-    line: usize,
-    context: &str,
-) {
-    output
-        .definitions
-        .push(project_item(file, kind, name, line, context));
-}
-
-fn push_reference(
-    file: &ScanFile,
-    output: &mut WorkerOutput,
-    kind: &str,
-    name: &str,
-    line: usize,
-    context: &str,
-) {
-    output
-        .references
-        .push(project_item(file, kind, name, line, context));
-}
-
-fn project_item(
-    file: &ScanFile,
-    kind: &str,
-    name: &str,
-    line: usize,
-    context: &str,
-) -> ProjectIndexItem {
-    ProjectIndexItem {
-        kind: kind.to_string(),
-        name: name.to_string(),
-        root: file.root.clone(),
-        root_role: file.root_role.clone(),
-        path: file.relative_path.clone(),
-        line,
-        context: context.to_string(),
-    }
-}
-
 fn sort_items(items: &mut [ProjectIndexItem]) {
     items.sort_by(|left, right| {
         (
@@ -519,20 +195,6 @@ fn sort_items(items: &mut [ProjectIndexItem]) {
                 &right.context,
             ))
     });
-}
-
-fn is_block_start(tokens: &[Token], index: usize) -> bool {
-    index + 2 < tokens.len()
-        && tokens[index].kind == TokenKind::Word
-        && tokens[index + 1].kind == TokenKind::Equals
-        && tokens[index + 2].kind == TokenKind::Open
-}
-
-fn is_assignment(tokens: &[Token], index: usize) -> bool {
-    index + 2 < tokens.len()
-        && tokens[index].kind == TokenKind::Word
-        && tokens[index + 1].kind == TokenKind::Equals
-        && matches!(tokens[index + 2].kind, TokenKind::Word | TokenKind::String)
 }
 
 fn should_index_file(relative_path: &str) -> bool {
@@ -623,167 +285,4 @@ fn worker_count(file_count: usize) -> usize {
         .map(usize::from)
         .unwrap_or(1)
         .clamp(1, file_count)
-}
-
-fn is_variable_key(key: &str) -> bool {
-    matches!(
-        key,
-        "set_variable"
-            | "set_temp_variable"
-            | "add_to_variable"
-            | "subtract_from_variable"
-            | "multiply_variable"
-            | "divide_variable"
-            | "modulo_variable"
-            | "clamp_variable"
-            | "round_variable"
-            | "check_variable"
-            | "has_variable"
-            | "clear_variable"
-    )
-}
-
-fn is_variable_name_field(key: &str) -> bool {
-    matches!(key, "var" | "variable" | "which")
-}
-
-fn variable_name_from_field<'a>(key: &'a str, value: &'a str) -> Option<&'a str> {
-    if is_variable_name_field(key) {
-        return Some(value);
-    }
-
-    (!is_variable_option_key(key) && is_script_identifier(key)).then_some(key)
-}
-
-fn is_variable_option_key(key: &str) -> bool {
-    matches!(
-        key,
-        "value"
-            | "min"
-            | "max"
-            | "add"
-            | "subtract"
-            | "multiply"
-            | "divide"
-            | "modulo"
-            | "tooltip"
-            | "days"
-            | "check_range_bounds"
-    )
-}
-
-fn is_script_identifier(value: &str) -> bool {
-    !value.is_empty()
-        && value
-            .chars()
-            .all(|character| character.is_ascii_alphanumeric() || matches!(character, '_' | '.'))
-}
-
-fn is_localisation_language_header(key: &str, rest: &str) -> bool {
-    key.starts_with("l_") && (rest.is_empty() || rest.starts_with('#'))
-}
-
-fn is_ignored_idea_block(key: &str) -> bool {
-    matches!(
-        key,
-        "ideas"
-            | "country"
-            | "political_advisor"
-            | "theorist"
-            | "army_chief"
-            | "navy_chief"
-            | "air_chief"
-            | "high_command"
-            | "designer"
-            | "industrial_concern"
-            | "materiel_manufacturer"
-            | "modifier"
-            | "allowed"
-            | "visible"
-            | "available"
-            | "allowed_civil_war"
-            | "cancel"
-            | "on_add"
-            | "on_remove"
-            | "traits"
-    )
-}
-
-#[cfg(test)]
-mod tests {
-    use std::fs;
-
-    use super::{ProjectIndexItem, ProjectIndexRequest, index_hoi4_project};
-    use crate::tools::{ScanRoot, test_support::unique_test_dir};
-
-    #[test]
-    fn indexes_hoi4_definitions_and_references() {
-        let root = unique_test_dir("project-index");
-        write_file(
-            &root,
-            "common/scripted_triggers/sample_triggers.txt",
-            "sample_has_system_ready = { has_country_flag = sample_system_ready check_variable = { sample_score > 0 } }\n",
-        );
-        write_file(
-            &root,
-            "common/scripted_effects/sample_effects.txt",
-            "sample_apply_system = { set_country_flag = sample_system_ready set_variable = { sample_score = 1 } }\n",
-        );
-        write_file(
-            &root,
-            "interface/sample_interface.gfx",
-            "spriteTypes = { spriteType = { name = \"GFX_sample_panel\" texturefile = \"gfx/interface/sample/panel.png\" } }\n",
-        );
-        write_file(
-            &root,
-            "interface/sample_interface.gui",
-            "guiTypes = { containerWindowType = { name = \"sample_panel_window\" background = { quadTextureSprite = \"GFX_sample_panel\" } } }\n",
-        );
-        write_file(
-            &root,
-            "localisation/simp_chinese/project_index_l_simp_chinese.yml",
-            "\u{feff}l_simp_chinese:\n sample_system_ready:0 \"系统\"\n",
-        );
-
-        let index = index_hoi4_project(ProjectIndexRequest {
-            roots: vec![ScanRoot {
-                path: root.to_string_lossy().to_string(),
-                role: Some("mod".to_string()),
-            }],
-            include_game_roots: Some(true),
-        })
-        .expect("index should build");
-
-        assert_eq!(index.scanned_roots, 1);
-        assert!(index.scanned_files >= 5);
-        assert_index_item(
-            &index.definitions,
-            "scripted_trigger",
-            "sample_has_system_ready",
-        );
-        assert_index_item(&index.definitions, "scripted_effect", "sample_apply_system");
-        assert_index_item(&index.definitions, "gfx_sprite", "GFX_sample_panel");
-        assert_index_item(&index.definitions, "gui_element", "sample_panel_window");
-        assert_index_item(&index.references, "country_flag", "sample_system_ready");
-        assert_index_item(&index.references, "variable", "sample_score");
-        assert_index_item(&index.references, "gfx_sprite", "GFX_sample_panel");
-
-        fs::remove_dir_all(root).expect("temp output should clean up");
-    }
-
-    fn write_file(root: &std::path::Path, relative_path: &str, content: &str) {
-        let path = root.join(relative_path);
-        if let Some(parent) = path.parent() {
-            fs::create_dir_all(parent).expect("fixture parent should be created");
-        }
-        fs::write(path, content).expect("fixture file should be written");
-    }
-
-    fn assert_index_item(items: &[ProjectIndexItem], kind: &str, name: &str) {
-        assert!(
-            items
-                .iter()
-                .any(|item| item.kind == kind && item.name == name)
-        );
-    }
 }
